run.py
#!/usr/bin/env python3
"""
UFC Fight Predictor - Main Entry Point

This script provides CLI commands to run different phases of the prediction system:
- ETL: Data collection and processing
- Training: Model training and evaluation
- Prediction: Generate predictions for upcoming fights
- Web: Launch the web interface

Usage:
    python run.py etl --historical          # Collect historical fight data
    python run.py etl --upcoming            # Collect upcoming fight data
    python run.py etl --update              # Update all data
    python run.py train                     # Train the prediction model
    python run.py train --tune              # Train with hyperparameter tuning
    python run.py predict                   # Generate predictions
    python run.py predict --event <name>    # Predict specific event
    python run.py web                       # Launch web interface
"""

import argparse
import logging
import yaml
import sys
from pathlib import Path

# Add src to path
sys.path.insert(0, str(Path(__file__).parent / 'src'))


def setup_logging(config: dict):
    """
    Set up logging configuration.

    Args:
        config: Configuration dictionary
    """
    logging.basicConfig(
        level=config.get('logging', {}).get('level', 'INFO'),
        format=config.get('logging', {}).get('format',
                                              '%(asctime)s - %(name)s - %(levelname)s - %(message)s'),
        handlers=[
            logging.StreamHandler(),
            logging.FileHandler(config.get('logging', {}).get('file', 'ufc_predictor.log'))
        ]
    )


def load_config(config_path: str = 'config.yaml') -> dict:
    """
    Load configuration from YAML file.

    Args:
        config_path: Path to configuration file

    Returns:
        Configuration dictionary
    """
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def run_etl_historical(config: dict):
    """
    Run ETL pipeline for historical fight data.

    TODO Phase 1:
    - Import ETL modules
    - Initialize scraper
    - Scrape historical data
    - Process and clean data
    - Save to processed directory
    """
    print("Running ETL for historical data...")
    print("TODO Phase 1: Implement historical data collection")
    # from src.etl import UFCDataScraper, DataProcessor
    # scraper = UFCDataScraper(config)
    # processor = DataProcessor(config)
    # ...


def run_etl_upcoming(config: dict):
    """
    Run ETL pipeline for upcoming fight data.

    TODO Phase 1:
    - Import ETL modules
    - Initialize scraper
    - Scrape upcoming events
    - Process and clean data
    - Save to processed directory
    """
    print("Running ETL for upcoming fights...")
    print("TODO Phase 1: Implement upcoming fights collection")


def run_etl_update(config: dict):
    """
    Run complete data update pipeline.

    TODO Phase 1:
    - Import DataUpdater
    - Run scheduled update
    - Update both historical and upcoming data
    """
    print("Running complete data update...")
    print("TODO Phase 1: Implement data update pipeline")
    # from src.etl import DataUpdater
    # updater = DataUpdater(config)
    # updater.run_scheduled_update()


def run_train(config: dict, tune_hyperparameters: bool = False):
    """
    Train the prediction model.

    Args:
        tune_hyperparameters: Whether to perform hyperparameter tuning

    TODO Phase 2:
    - Load processed data
    - Create features
    - Train model
    - Evaluate performance
    - Save model
    """
    print("Training prediction model...")
    if tune_hyperparameters:
        print("With hyperparameter tuning enabled")
    print("TODO Phase 2: Implement model training")
    # from src.features import FeatureEngineer
    # from src.models import ModelTrainer
    # ...


def run_predict(config: dict, event_name: str = None):
    """
    Generate predictions for upcoming fights.

    Args:
        event_name: Optional specific event to predict

    TODO Phase 3:
    - Load trained model
    - Load upcoming fights
    - Create features
    - Generate predictions
    - Save predictions
    """
    print("Generating predictions...")
    if event_name:
        print(f"For event: {event_name}")
    print("TODO Phase 3: Implement prediction generation")
    # from src.models import FightPredictor
    # from src.features import FeatureEngineer
    # ...


def run_web(config: dict):
    """
    Launch the web interface.

    TODO Phase 4:
    - Import web app
    - Load configuration
    - Start Flask server
    """
    print("Launching web interface...")
    print(f"Server will run on {config['web']['host']}:{config['web']['port']}")
    print("TODO Phase 4: Implement web interface")
    # from src.web import create_app
    # app = create_app(config)
    # app.run(host=config['web']['host'], port=config['web']['port'], debug=config['web']['debug'])


def main():
    """
    Main entry point with argument parsing.
    """
    parser = argparse.ArgumentParser(
        description='UFC Fight Predictor - ML-Powered Fight Predictions',
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog=__doc__
    )

    subparsers = parser.add_subparsers(dest='command', help='Command to run')

    # ETL commands
    etl_parser = subparsers.add_parser('etl', help='Run ETL pipeline')
    etl_group = etl_parser.add_mutually_exclusive_group(required=True)
    etl_group.add_argument('--historical', action='store_true',
                           help='Collect historical fight data')
    etl_group.add_argument('--upcoming', action='store_true',
                           help='Collect upcoming fight data')
    etl_group.add_argument('--update', action='store_true',
                           help='Update all data')

    # Training commands
    train_parser = subparsers.add_parser('train', help='Train prediction model')
    train_parser.add_argument('--tune', action='store_true',
                              help='Perform hyperparameter tuning')

    # Prediction commands
    predict_parser = subparsers.add_parser('predict', help='Generate predictions')
    predict_parser.add_argument('--event', type=str,
                                help='Predict specific event')

    # Web interface command
    subparsers.add_parser('web', help='Launch web interface')

    # Config argument for all commands
    parser.add_argument('--config', type=str, default='config.yaml',
                        help='Path to configuration file')

    args = parser.parse_args()

    # Load configuration
    try:
        config = load_config(args.config)
        setup_logging(config)
    except FileNotFoundError:
        print(f"Error: Configuration file '{args.config}' not found")
        return 1
    except yaml.YAMLError as e:
        print(f"Error parsing configuration file: {e}")
        return 1

    # Route to appropriate function
    try:
        if args.command == 'etl':
            if args.historical:
                run_etl_historical(config)
            elif args.upcoming:
                run_etl_upcoming(config)
            elif args.update:
                run_etl_update(config)
        elif args.command == 'train':
            run_train(config, tune_hyperparameters=args.tune)
        elif args.command == 'predict':
            run_predict(config, event_name=args.event)
        elif args.command == 'web':
            run_web(config)
        else:
            parser.print_help()
            return 1

    except Exception as e:
        logging.error(f"Error: {e}", exc_info=True)
        return 1

    return 0


if __name__ == '__main__':
    sys.exit(main())
