requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping & API
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Web Framework
flask>=3.0.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
